requirements.txt
Flask
Werkzeug
shortuuid
flask-socketio
flask-login
flask-sqlalchemy
sentry-sdk
blinker
psycopg2-binary
eventlet
gunicorn
rq
pillow
